Only delete subjects without other sessions, add backoff for 502/504 [GEAR-491]

+++ run.py
@@ -11,6 +11,7 @@ import tempfile
 import time
 import zipfile
 
+import backoff
 from pathlib import Path
 from pprint import pprint as pp
 
@@ -26,9 +27,18 @@ log = logging.getLogger('[GRP 9]:')
 log.setLevel(logging.INFO)
 
 
+def false_if_exc_is_timeout(exception):
+    if exception.status in [504, 502]:
+        return False
+    return True
+
+
 ###############################################################################
 # LOCAL FUNCTION DEFINITIONS
 
+
+@backoff.on_exception(backoff.expo, flywheel.rest.ApiException,
+                      max_time=300, giveup=false_if_exc_is_timeout)
 def _find_or_create_subject(fw, session, project, subject_code):
     # Try to find if a subject with that code already exists in the project
     query_code = quote_numeric_string(subject_code)
@@ -68,6 +78,8 @@ def _find_or_create_subject(fw, session, project, subject_code):
     return subject
 
 
+@backoff.on_exception(backoff.expo, flywheel.rest.ApiException,
+                      max_time=300, giveup=false_if_exc_is_timeout)
 def _archive_session(fw, session, archive_project):
     """Move session to archive project
 
@@ -284,6 +296,8 @@ def _modify_dicom_archive(dicom_file_path, update_keys, flywheel_dicom_header, d
     return modified_dicom_file_path
 
 
+@backoff.on_exception(backoff.expo, flywheel.rest.ApiException,
+                      max_time=300, giveup=false_if_exc_is_timeout)
 def validate_classification(fw, f_modality, f_classification, f_name):
     """Make sure classification is valid under the modality schema.
 
@@ -300,6 +314,13 @@ def validate_classification(fw, f_modality, f_classification, f_name):
 
     valid_for_modality = True
     classification_schema = dict()
+    if not f_modality:
+        log_msg = (
+            f'file {f_name} does not have a modality. It will be assumed that'
+            f' classification {f_classification} is valid'
+        )
+        log.info(log_msg)
+        return valid_for_modality
     try:
         classification_schema = fw.get_modality(f_modality)
 
@@ -336,6 +357,8 @@ def validate_classification(fw, f_modality, f_classification, f_name):
     return valid_for_modality
 
 
+@backoff.on_exception(backoff.expo, flywheel.rest.ApiException,
+                      max_time=300, giveup=false_if_exc_is_timeout)
 def _export_files(fw, acquisition, export_acquisition, session, subject, project, config):
     """Export acquisition files to the exported acquisiton.
 
@@ -410,6 +433,40 @@ def _export_files(fw, acquisition, export_acquisition, session, subject, project
             export_acquisition.reload()
 
 
+@backoff.on_exception(backoff.expo, flywheel.rest.ApiException,
+                      max_time=300, giveup=false_if_exc_is_timeout)
+def ok_to_delete_subject(fw_client, subject_dict, session_dict_list):
+    """
+    Determine whether it is appropriate to delete the subject, based on whether
+        the subject has sessions outside of those created by this gear
+    Args:
+        fw_client (flywheel.Client): an instance of the flywheel client
+        subject_dict (dict): a dictionary that contains key 'id' with a str
+            value equal to a flywheel subject id
+        session_dict_list (list): a list of dicts that contain key 'id' with a
+            str value equal to a flywheel session id
+
+    Returns:
+        bool: whether the subject can safely be deleted
+    """
+
+    session_ids = [sess['id'] for sess in session_dict_list]
+    try:
+        subject_obj = fw_client.get_subject(subject_dict.get('id'))
+
+        for session in subject_obj.sessions.iter():
+            if session.id not in session_ids:
+                return False
+    except flywheel.rest.ApiException as exc:
+        if exc.status in [403, 404]:
+            return False
+        else:
+            raise exc
+    return True
+
+
+@backoff.on_exception(backoff.expo, flywheel.rest.ApiException,
+                      max_time=300, giveup=false_if_exc_is_timeout)
 def _cleanup(fw, creatio):
     """
     In the case of a failure, cleanup all containers that were created.
@@ -419,22 +476,59 @@ def _cleanup(fw, creatio):
     if acquisitions:
         log.info("Deleting {} acquisition containers".format(len(acquisitions)))
         for a in acquisitions:
-            log.debug(a)
-            fw.delete_acquisition(a['id'])
+            try:
+                log.debug(a)
+                fw.delete_acquisition(a['id'])
+            except flywheel.rest.ApiException as exc:
+                if exc.status in [502, 504]:
+                    raise exc
+                else:
+                    a_id = a['id']
+                    error_msg = (
+                        'Exception encountered when attempting to delete'
+                        f' acquisition {a_id}.'
+                    )
+                    log.error(error_msg, exc_info=True)
+                    continue
 
     sessions = [ x for x in creatio if x['container'] == "session" ]
     if sessions:
         log.info("Deleting {} session containers".format(len(sessions)))
         for s in sessions:
             log.debug(s)
-            fw.delete_session(s['id'])
+            try:
+                fw.delete_session(s['id'])
+            except flywheel.rest.ApiException as exc:
+                if exc.status in [502, 504]:
+                    raise exc
+                else:
+                    s_id = s['id']
+                    error_msg = (
+                        'Exception encountered when attempting to delete'
+                        f' session {s_id}.'
+                    )
+                    log.error(error_msg, exc_info=True)
+                    continue
 
     subjects = [ x for x in creatio if x['container'] == "subject" and x['new'] == True ]
     if subjects:
         log.info("Deleting {} subject containers".format(len(subjects)))
         for s in subjects:
-            log.debug(s)
-            fw.delete_subject(s['id'])
+            if ok_to_delete_subject(fw, s, sessions):
+                log.debug('Deleting subject %s', s.get('id'))
+                try:
+                    fw.delete_subject(s['id'])
+                except flywheel.rest.ApiException as exc:
+                    if exc.status in [502, 504]:
+                        raise exc
+                    else:
+                        s_id = s['id']
+                        error_msg = (
+                            'Exception encountered when attempting to delete'
+                            f' subject {s_id}.'
+                        )
+                        log.error(error_msg, exc_info=True)
+                        continue
 
 
 def main(context):
@@ -739,6 +833,7 @@ def get_patientsex_from_subject(subject):
     else:
         return ''
 ###############################################################################
+
 
 if __name__ == '__main__':
     with flywheel.GearContext() as context:
